requirements.txt
google-generativeai
requests
beautifulsoup4
pandas
python-dotenv
beautifulsoup4
spacy